review update
reformat, copyright, moved LockCriticalSection to shared folder
